Task 6

+++ 0x00-hello_world/6-size.c
@@ -8,11 +8,11 @@
 int main(void)
 {
 
-printf("char size: %lu byte(s)\n", sizeof(char));
-printf("int size: %lu byte(s)\n", sizeof(int));
-printf("long int size: %lu byte(s)\n", sizeof(long int));
-printf("long long int size: %lu byte(s)\n", sizeof(long long int));
-printf("float size: %lu byte(s)\n", sizeof(float));
+printf("char size: %lu byte(s)\n", (unsigned long)sizeof(char));
+printf("int size: %lu byte(s)\n", (unsigned long)sizeof(int));
+printf("long int size: %lu byte(s)\n", (unsigned long)sizeof(long int));
+printf("long long int size: %lu byte(s)\n", (unsigned long)sizeof(long long int));
+printf("float size: %lu byte(s)\n", (unsigned long)sizeof(float));
 
 	return (0);
